Refactor CLI command model into modules
